backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-restful==0.3.9
flask-sqlalchemy==3.0.3
werkzeug==2.2.3
python-dotenv==1.0.0
requests==2.28.2
pytest==7.3.1
black==23.3.0
flake8==6.0.0
supabase==2.0.2
pyjwt==2.6.0
SQLAlchemy==1.4.46
aniso8601==9.0.1
colorama==0.4.6
uuid==1.30
huggingface_hub==0.21.4